+++ CHANGELOG.md
@@ -13,6 +13,7 @@ Release: dd.mm.yyyy
 ### Improvements
 
 - Callback for key is now called repeatedly when key is held down ([#119](https://github.com/kasper/phoenix/issues/119)).
+- Performance improvements.
 
 ### Bug Fixes